# default.openapi .json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Default module",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/suppliers": {
      "get": {
        "summary": "get-all",
        "deprecated": false,
        "description": "",
        "tags": [
          "suppliers"
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "",
            "required": false,
            "example": 1,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "",
            "required": false,
            "example": 1,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "seach",
            "in": "query",
            "description": "",
            "required": false,
            "example": "name",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                },
                "example": {
                  "data": [
                    {
                      "supplier_id": 1,
                      "supplier_name": "CP1",
                      "phone": "0999999911"
                    },
                    {
                      "supplier_id": 2,
                      "supplier_name": "CP2",
                      "phone": "0999999922"
                    },
                    {
                      "supplier_id": 3,
                      "supplier_name": "CP3",
                      "phone": "0999999933"
                    }
                  ],
                  "pagination": {
                    "page": 1,
                    "limit": 10,
                    "total": 3,
                    "total_pages": 1
                  }
                }
              }
            },
            "headers": {}
          },
          "400": {
            "$ref": "#/components/responses/400 Bad Request",
            "description": ""
          },
          "401": {
            "$ref": "#/components/responses/401 Unauthorized",
            "description": ""
          },
          "500": {
            "$ref": "#/components/responses/500 Internal Server Error",
            "description": ""
          }
        },
        "security": []
      },
      "post": {
        "summary": "create",
        "deprecated": false,
        "description": "",
        "tags": [
          "suppliers"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {}
              },
              "example": {
                "supplier_name": "ABC Farm",
                "phone": "0812345678",
                "address": "Bangkok"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                },
                "example": {
                  "message": "Supplier created successfully"
                }
              }
            },
            "headers": {}
          },
          "400": {
            "$ref": "#/components/responses/400 Bad Request",
            "description": ""
          },
          "401": {
            "$ref": "#/components/responses/401 Unauthorized",
            "description": ""
          },
          "409": {
            "$ref": "#/components/responses/409 Conflict",
            "description": ""
          },
          "500": {
            "$ref": "#/components/responses/500 Internal Server Error",
            "description": ""
          }
        },
        "security": []
      }
    },
    "/suppliers/{id}": {
      "get": {
        "summary": "get-one",
        "deprecated": false,
        "description": "",
        "tags": [
          "suppliers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                },
                "example": {
                  "supplier_id": 1,
                  "supplier_name": "CP",
                  "phone": "0999999999",
                  "address": "24 ถนนตีทอง เเขวงป้อนปราบ เขตบางใหญ กรุงเทพ 10200"
                }
              }
            },
            "headers": {}
          },
          "400": {
            "$ref": "#/components/responses/400 Bad Request",
            "description": ""
          },
          "401": {
            "$ref": "#/components/responses/401 Unauthorized",
            "description": ""
          },
          "404": {
            "$ref": "#/components/responses/404 Not Found",
            "description": ""
          },
          "500": {
            "$ref": "#/components/responses/500 Internal Server Error",
            "description": ""
          }
        },
        "security": []
      },
      "delete": {
        "summary": "delete",
        "deprecated": false,
        "description": "",
        "tags": [
          "suppliers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                },
                "example": {
                  "message": "Supplier deleted successfully"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "put": {
        "summary": "update",
        "deprecated": false,
        "description": "",
        "tags": [
          "suppliers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {}
              },
              "example": {
                "supplier_name": "ABC Farm Updated",
                "phone": "0899999999",
                "address": "Chiang Mai"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                },
                "example": {
                  "message": "Supplier updated",
                  "data": {
                    "supplier_id": 1,
                    "supplier_name": "ABC Farm Updated",
                    "phone": "0899999999",
                    "address": "Chiang Mai",
                    "status": "active"
                  }
                }
              }
            },
            "headers": {}
          },
          "400": {
            "$ref": "#/components/responses/400 Bad Request",
            "description": ""
          },
          "401": {
            "$ref": "#/components/responses/401 Unauthorized",
            "description": ""
          },
          "404": {
            "$ref": "#/components/responses/404 Not Found",
            "description": ""
          },
          "500": {
            "$ref": "#/components/responses/500 Internal Server Error",
            "description": ""
          }
        },
        "security": []
      }
    },
    "/suppliers/inactive/{id}": {
      "put": {
        "summary": "activeUpdate",
        "deprecated": false,
        "description": "",
        "tags": [
          "suppliers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {}
              },
              "example": {
                "status": "inactive"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                },
                "example": {
                  "message": "Supplier updated",
                  "data": {
                    "supplier_id": 1,
                    "status": "inactive"
                  }
                }
              }
            },
            "headers": {}
          },
          "400": {
            "$ref": "#/components/responses/400 Bad Request",
            "description": ""
          },
          "401": {
            "$ref": "#/components/responses/401 Unauthorized",
            "description": ""
          },
          "404": {
            "$ref": "#/components/responses/404 Not Found",
            "description": ""
          },
          "500": {
            "$ref": "#/components/responses/500 Internal Server Error",
            "description": ""
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {}
  },
  "servers": [],
  "security": []
}